Fixes styling on premium & bug without PIN
If the user did not have a PIN set up, the script broke because it could not find the PIN table.

When the user does not have Premium, we need to retain the functionality that broke the stock ticker on premium, or else the stocks table can't be full width.