Fixes bugs with Stock Highlighter script (#8)
* Reworks table style cleanup to address #6
When on the portfolio page logged into an account with Neopets Premium, the marquee ticker was not visible. This seemeed to be caused by some code that was removing a div (ads?) from the page.

* Fixes styling on premium & bug without PIN

If the user did not have a PIN set up, the script broke because it could not find the PIN table.

When the user does not have Premium, we need to retain the functionality that broke the stock ticker on premium, or else the stocks table can't be full width.